+++ .expeditor/scripts/end_to_end/SupervisorRunner.cs
@@ -5,7 +5,7 @@ using System.IO;
 public class SupervisorRunner {
     private String logPath = null;
 
-    public Process Run(string logPath) {
+    public Process Run(string logPath, string[] args) {
         this.logPath = logPath;
         var proc = new Process();
         proc.StartInfo.UseShellExecute = false;
@@ -13,7 +13,7 @@ public class SupervisorRunner {
         proc.StartInfo.RedirectStandardOutput = true;
         proc.StartInfo.RedirectStandardError = true;
         proc.StartInfo.FileName = "hab";
-        proc.StartInfo.Arguments = "sup run --no-color";
+        proc.StartInfo.Arguments = "sup run --no-color " + String.Join(" ", args);
         proc.StartInfo.EnvironmentVariables["HAB_NOCOLORING"] = "1";
         proc.OutputDataReceived += new DataReceivedEventHandler(SupOutputHandler);
         proc.ErrorDataReceived += new DataReceivedEventHandler(SupOutputHandler);